Refactor feedback-panel.tsx: extract detail panel and finding row
